// src/vendor/rrweb/rrweb.d.ts
export declare type playerConfig = {
  speed: number;
  maxSpeed: number;
  root: Element;
  loadTimeout: number;
  skipInactive: boolean;
  inactivePeriodThreshold: number;
  showWarning: boolean;
  showDebug: boolean;
  blockClass: string;
  liveMode: boolean;
  insertStyleRules: string[];
  triggerFocus: boolean;
  UNSAFE_replayCanvas: boolean;
  pauseAnimation?: boolean;
  mouseTail:
    | boolean
    | {
        duration?: number;
        lineCap?: string;
        lineWidth?: number;
        strokeStyle?: string;
      };
  unpackFn?: UnpackFn;
  useVirtualDom: boolean;
  logger: {
    log: (...args: Parameters<typeof console.log>) => void;
    warn: (...args: Parameters<typeof console.warn>) => void;
  };
  plugins?: ReplayPlugin[];
};
// ...
export declare class Replayer {
  wrapper: HTMLDivElement;
  iframe: HTMLIFrameElement;
  service: ReturnType<typeof createPlayerService>;
  speedService: ReturnType<typeof createSpeedService>;
  get timer(): Timer;
  config: playerConfig;
  usingVirtualDom: boolean;
  virtualDom: RRDocument;
  private mouse;
  private mouseTail;
  private tailPositions;
  private emitter;
  private nextUserInteractionEvent;
  private legacy_missingNodeRetryMap;
  private cache;
  private imageMap;
  private canvasEventMap;
  private mirror;
  private styleMirror;
  private mediaManager;
  private firstFullSnapshot;
  private newDocumentQueue;
  private mousePos;
  private touchActive;
  private lastMouseDownEvent;
  private lastHoveredRootNode;
  private lastSelectionData;
  private constructedStyleMutations;
  private adoptedStyleSheets;
  constructor(events: Array<eventWithTime | string>, config?: Partial<playerConfig>);
  on(event: string, handler: Handler): this;
  off(event: string, handler: Handler): this;
  setConfig(config: Partial<playerConfig>): void;
  getMetaData(): playerMetaData;
  getCurrentTime(): number;
  getTimeOffset(): number;
  getMirror(): Mirror;
  play(timeOffset?: number): void;
  pause(timeOffset?: number): void;
  resume(timeOffset?: number): void;
  destroy(): void;
  startLive(baselineTime?: number): void;
  addEvent(rawEvent: eventWithTime | string): void;
  enableInteract(): void;
  disableInteract(): void;
  resetCache(): void;
  private setupDom;
  private handleResize;
  private applyEventsSynchronously;
  private getCastFn;
  private rebuildFullSnapshot;
  private insertStyleRules;
  private attachDocumentToIframe;
  private collectIframeAndAttachDocument;
  private waitForStylesheetLoad;
  private preloadAllImages;
  private preloadImages;
  private deserializeAndPreloadCanvasEvents;
  private applyIncremental;
  private applyMutation;
  private applyScroll;
  private applyInput;
  private applySelection;
  private applyStyleSheetMutation;
  private applyStyleSheetRule;
  private applyStyleDeclaration;
  private applyAdoptedStyleSheet;
  private legacy_resolveMissingNode;
  private moveAndHover;
  private drawMouseTail;
  private hoverElements;
  private isUserInteraction;
  private backToNormal;
  private warnNodeNotFound;
  private warnCanvasMutationFailed;
  private debugNodeNotFound;
  private warn;
  private debug;
}

export { ReplayerEvents };

export declare type ReplayPlugin = {
  handler?: (
    event: eventWithTime,
    isSync: boolean,
    context: {
      replayer: Replayer;
    }
  ) => void;
  onBuild?: (
    node: Node | RRNode,
    context: {
      id: number;
      replayer: Replayer;
    }
  ) => void;
  getMirror?: (mirrors: { nodeMirror: Mirror }) => void;
};
// ...
declare class Timer {
  timeOffset: number;
  speed: number;
  private actions;
  private raf;
  private lastTimestamp;
  constructor(
    actions: actionWithDelay[] | undefined,
    config: {
      speed: number;
    }
  );
  addAction(action: actionWithDelay): void;
  start(): void;
  private rafCheck;
  clear(): void;
  setSpeed(speed: number): void;
  isActive(): boolean;
  private findActionIndex;
}